fix docker setup and server configuration

+++ database/migrations/2025_11_03_184011_change_subject_id_to_topic_id_in_chapters_table.php
@@ -10,14 +10,16 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up()
-    {
-        Schema::table('chapters', function (Blueprint $table) {
-            $table->dropForeign(['subject_id']);
-            $table->dropColumn('subject_id');
-
-            $table->foreignId('topic_id')->after('id')->constrained();
-        });
-    }
+	{
+   	 Schema::table('chapters', function (Blueprint $table) {
+        if (Schema::hasColumn('chapters', 'subject_id')) {
+       		     $table->dropColumn('subject_id');
+       		 }
+        if (!Schema::hasColumn('chapters', 'topic_id')) {
+    	       	     $table->foreignId('topic_id')->after('id')->constrained();
+        	}
+	    });
+	}
 
     public function down()
     {
